fix SW_HEADPHONE_INSERT is not send when first plug in
The fixed SW_HEADPHONE_INSERT is not send when first plug in.
input_report_switch function is only allow to send the value of
SW_HEADPHONE_INSERT when it changes.
The first time value of SW_HEADPHONE_INSERT is 0.
so need to change to value to 1 to solve the bug.